JetBot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
amqp==5.3.1
anyio==4.10.0
asn1crypto==1.5.1
blinker==1.9.0
boto3==1.40.33
botocore==1.40.33
cachelib==0.13.0
agiref==3.9.1
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
channels==4.3.1
channels_redis==4.3.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.7.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
kombu==5.5.4
msgpack==1.1.1
multidict==6.6.4
packaging==25.0
prompt_toolkit==3.0.52
propcache==0.3.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
PyMuPDF==1.26.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.48.0
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.35.0
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
whitenoise==6.10.0
yarl==1.20.1
